Revert "Frontend fix"
