--- src/back-end/CryptEx/CryptExApi/Models/DTO/IbanDto.cs
@@ -1,15 +1,0 @@
-﻿using CryptExApi.Validators;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace CryptExApi.Models.DTO
-{
-    public class IbanDto
-    {
-        [Required, IbanValidator]
-        public string Iban { get; set; }
-    }
-}
