+++ src/back-end/CryptEx/CryptExApi/Models/Database/AppUser.cs
@@ -23,9 +23,7 @@ namespace CryptExApi.Models.Database
 
         public string PreferedCurrency { get; set; } = "usd";
 
-        public UserAddress Address { get; set; }
-
-        public BankAccount BankAccount { get; set; }
+        public List<BankAccount> BankAccounts { get; set; }
     }
 
     public class AppRole : IdentityRole<Guid>
