+++ src/back-end/CryptEx/CryptExApi/Models/Database/UserAddress.cs
@@ -21,9 +21,5 @@ namespace CryptExApi.Models.Database
         public Guid CountryId { get; set; }
 
         public Country Country { get; set; }
-
-        public Guid UserId { get; set; }
-
-        public AppUser User { get; set; }
     }
 }
