+++ src/front-end/CryptExFrontEnd/package.json
@@ -21,7 +21,6 @@
     "@angular/platform-browser": "~11.2.12",
     "@angular/platform-browser-dynamic": "~11.2.12",
     "@angular/router": "~11.2.12",
-    "lightweight-charts": "^3.3.0",
     "@auth0/angular-jwt": "^5.0.2",
     "@microsoft/signalr": "^5.0.6",
     "@ngx-translate/core": "^13.0.0",
@@ -35,16 +34,13 @@
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^11.1.1",
     "@angular-devkit/build-angular": "~0.1102.11",
     "@angular/cli": "~11.2.11",
     "@angular/compiler-cli": "~11.2.12",
-    "@tailwindcss/forms": "^0.3.2",
-    "@angular/localize": "^11.2.12",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "autoprefixer": "^10.2.5",
+    "autoprefixer": "^10.0.2",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
@@ -54,12 +50,13 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "ngx-build-plus": "^10.1.1",
-    "postcss": "^8.2.15",
+    "ngx-tailwind": "^1.1.0",
+    "postcss": "^8.1.7",
     "postcss-import": "^13.0.0",
-    "postcss-loader": "^4.3.0",
+    "postcss-loader": "^4.0.4",
     "postcss-scss": "^3.0.4",
     "protractor": "~7.0.0",
-    "tailwindcss": "^2.1.2",
+    "tailwindcss": "^2.0.1",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
     "typescript": "~4.0.2"
