+++ src/front-end/CryptExFrontEnd/src/app/main/Components/contact/contact.component.html
@@ -6,17 +6,17 @@
       <div class="bg-gray-50 py-16 px-4 sm:px-6 lg:col-span-2 lg:px-8 lg:py-24 xl:pr-12">
         <div class="max-w-lg mx-auto">
           <h2 class="text-2xl font-extrabold tracking-tight text-gray-900 sm:text-3xl">
-            {{ 'Contact.ContactUs' | translate }}
+            Contact Us
           </h2>
           <p class="mt-3 text-lg leading-6 text-gray-500">
-            {{ 'Contact.Description' | translate }}
+            Our team is ready to hear your questions
           </p>
           <dl class="mt-8 text-base text-gray-500">
             <div>
-              <dt class="sr-only">{{ 'Contact.Address.Title' | translate }}</dt>
+              <dt class="sr-only">Postal address</dt>
               <dd>
-                <p>{{ 'Contact.Address.Line1' | translate }}</p>
-                <p>{{ 'Contact.Address.Line2' | translate }}</p>
+                <p>Rue de Genève 63</p>
+                <p>1002 Lausanne</p>
               </dd>
             </div>
             <div class="mt-6">
