+++ src/front-end/CryptExFrontEnd/src/app/main/Components/header/header.component.html
@@ -22,7 +22,7 @@
 
               <a routerLink="buySell" routerLinkActive="bg-gray-900" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ 'Header.NavBar.BuySell' | translate}}</a>
 
-              <a routerLink="wallets" routerLinkActive="bg-gray-900" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ (IsAuthenticated() ? 'Header.NavBar.MyWallets' : 'Header.NavBar.Wallets') | translate }}</a>
+              <a routerLink="wallets" routerLinkActive="bg-gray-900" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ 'Header.NavBar.MyWallets' | translate}}</a>
 
               <a routerLink="deposit-withdraw" routerLinkActive="bg-gray-900" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ 'Header.NavBar.DepositWithdraw' | translate}}</a>
 
@@ -69,11 +69,11 @@
         <!-- Current: "bg-gray-900 text-white", Default: "text-gray-300 hover:bg-gray-700 hover:text-white" -->
         <a routerLink="home"    routerLinkActive="bg-gray-900 text-white" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ 'Header.NavBar.Home' | translate }}</a>
 
-        <a routerLink="buySell" routerLinkActive="bg-gray-900 text-white" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ 'Header.NavBar.BuySell' | translate }}</a>
+        <a routerLink="buySell" routerLinkActive="bg-gray-900 text-white" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ 'Header.NavBar.BuySell' | translate}}</a>
 
-        <a routerLink="wallets" routerLinkActive="bg-gray-900 text-white" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ (IsAuthenticated() ? 'Header.NavBar.MyWallets' : 'Header.NavBar.Wallets') | translate }}</a>
+        <a routerLink="wallets" routerLinkActive="bg-gray-900 text-white" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ 'Header.NavBar.MyWallets' | translate}}</a>
 
-        <a routerLink="deposit-withdraw" routerLinkActive="bg-gray-900 text-white" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ 'Header.NavBar.DepositWithdraw' | translate }}</a>
+        <a routerLink="deposit-withdraw" routerLinkActive="bg-gray-900 text-white" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ 'Header.NavBar.DepositWithdraw' | translate}}</a>
 
         <a routerLink="contact" routerLinkActive="bg-gray-900 text-white" class="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium">{{ 'Header.NavBar.Contact' | translate }}</a>
       </div>
