+++ src/front-end/CryptExFrontEnd/src/app/user/Components/wallets/wallets.component.html
@@ -1,0 +1,1 @@
+<p>wallets works!</p>
