+++ src/front-end/CryptExFrontEnd/src/app/user/Components/wallets/wallets.component.spec.ts
