+++ src/front-end/CryptExFrontEnd/src/app/user/Components/wallets/wallets.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-wallets',
+  templateUrl: './wallets.component.html',
+  styleUrls: ['./wallets.component.scss']
+})
+export class WalletsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
