+++ src/front-end/CryptExFrontEnd/src/app/user/user.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { WalletsComponent } from './Components/wallets/wallets.component';
 import { MyAccountComponent } from './Components/my-account/my-account.component';
 import { UserRouting } from './user-routing';
 import { FormsModule } from '@angular/forms';
@@ -10,6 +11,7 @@ import { TranslateModule } from '@ngx-translate/core';
 
 @NgModule({
   declarations: [
+    WalletsComponent,
     MyAccountComponent
   ],
   imports: [
