--- src/front-end/CryptExFrontEnd/src/app/wallet/components/asset-card/asset-card.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { AssetCardComponent } from './asset-card.component';
-
-describe('AssetCardComponent', () => {
-  let component: AssetCardComponent;
-  let fixture: ComponentFixture<AssetCardComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ AssetCardComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(AssetCardComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
