--- src/front-end/CryptExFrontEnd/src/app/wallet/components/asset-card/asset-card.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-asset-card',
-  templateUrl: './asset-card.component.html',
-  styleUrls: ['./asset-card.component.scss']
-})
-export class AssetCardComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
