--- src/front-end/CryptExFrontEnd/src/app/wallet/components/wallets/wallets.component.scss
@@ -1,19 +1,0 @@
-
-.floating-tooltip-2 {
-	width: 96px;
-	height: 80px;
-	position: absolute;
-	display: none;
-	padding: 8px;
-	box-sizing: border-box;
-	font-size: 12px;
-	color: #131722;
-	background-color: rgba(255, 255, 255, 1);
-	text-align: left;
-	z-index: 1000;
-	top: 12px;
-	left: 12px;
-	pointer-events: none;
-	border: 1px solid rgba(255, 70, 70, 1);
-	border-radius: 2px;
-}
